fully split apart remaining storage folder operations
