+++ modules/host/contractmanager/storagefolderremove_test.go
